Rearrange directories and headers, create directory tablebase, header files persistence
