+++ src/process_pgn/read_pgn_data.cpp
@@ -1,7 +1,8 @@
 #include "process_pgn/read_pgn_data.hpp"
 #include "process_pgn/pgn_game.hpp"
-#include "threadpool.hpp"
+#include "threadpool/threadpool.hpp"
 #include "tablebase/master_tablebase.hpp"
+#include "tablebase/persistence.hpp"
 #include <chrono>
 #include <utility>
 #include <filesystem>
